Show paid subscription needed dialog when unapproved user tries to send a mail, fix #2781
We don't totally block the user from opening up a mail editor dialog, because they should still be able to send an email to sales@tutao.de. Instead we just show the subscription needed dialog when they try to open, and then again when they try to send
